web(ProjectPicker): in-place "Connect a daemon" + back-to-workspace
The picker's "Connect a new daemon" button deep-linked into /onboarding,
throwing users into the full wizard. Three fixes:

1. Connect in-place: new ConnectDaemonModal offers both managed-cloud
   (CreateDaemon/resume) and self-hosted paths without leaving the picker.
2. Self-hosted: extract SelfHostedDaemonConnect (token via CreateDaemonToken
   + `reliant daemon start --token` install steps) from onboarding's
   ComputeStep so both surfaces share one component. ComputeStep now consumes
   it (dead download/OS-detect/PAT code removed).
3. Back-to-app: "Back to <project>" affordance re-selects the most recent
   project to restore the chat/workspace shell; hidden on first-run.
